requirements.txt
alpha_vantage==3.0.0
beautifulsoup4==4.12.3
cvxpy==1.6.0
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
plotly==5.24.1
pyportfolioopt==1.5.6
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
yfinance==0.2.51
mysql-connector-python==9.1.0
bcrypt==3.2.0
streamlit==1.41.1

